+++ HISTORY.rst
@@ -11,7 +11,7 @@ API Changes (Backward-Compatible)
   that controls what encoding is used (if any) to decode the headers from bytes
   to unicode. This defaults to UTF-8 for backward compatibility. To disable the
   decode and use bytes exclusively, set the field to False, None, or the empty
-  string.
+  string. This affects all headers, including those pushed by servers.
 - Bumped the minimum version of HPACK allowed from 2.0 to 2.1.
 
 2.2.3 (2016-04-13)