+++ HISTORY.rst
@@ -13,6 +13,7 @@ API Changes (Backward-Compatible)
   decode and use bytes exclusively, set the field to False, None, or the empty
   string. This affects all headers, including those pushed by servers.
 - Bumped the minimum version of HPACK allowed from 2.0 to 2.1.
+- Added support for advertising RFC 7838 Alternative services.
 
 2.2.3 (2016-04-13)
 ------------------