+++ HISTORY.rst
@@ -14,6 +14,11 @@ API Changes (Backward-Compatible)
   string. This affects all headers, including those pushed by servers.
 - Bumped the minimum version of HPACK allowed from 2.0 to 2.2.
 - Added support for advertising RFC 7838 Alternative services.
+- Allowed users to provide ``hpack.HeaderTuple`` and
+  ``hpack.NeverIndexedHeaderTuple`` objects to all methods that send headers.
+- Changed all events that carry headers to emit ``hpack.HeaderTuple`` and
+  ``hpack.NeverIndexedHeaderTuple`` instead of plain tuples. This allows users
+  to maintain header indexing state.
 
 Bugfixes
 ~~~~~~~~