+++ HISTORY.rst
@@ -1,6 +1,17 @@
 Release History
 ===============
 
+2.1.5 (2016-04-25)
+------------------
+
+*Final 2.1.X release*
+
+Bugfixes
+~~~~~~~~
+
+- Correctly forbid pseudo-headers that were not defined in RFC 7540.
+- Ignore AltSvc frames, rather than exploding when receiving them.
+
 2.1.4 (2016-04-13)
 ------------------